refactor(whatsapp): modularize WhatsApp parameter management and queueing logic
Extract parameter handling into `ManagesWhatsAppParameters` and queueing logic into `HasQueueing` traits, ensuring better reusability and maintainability. Replace fluent setters with DTO conversion, tighten type safety, and introduce fallback handling for invalid WhatsApp types.
